refactor(fuzz): extract result-envelope persistence into core service
The fuzz report command owned result-envelope persistence directly —
opening the observation store, writing the envelope file, recording the
artifact, and building evidence refs — tripping the thin_command_adapter
audit detector (direct filesystem mutation + run artifact persistence).

Move the envelope-persistence cluster (persist_fuzz_result_envelope,
persist_fuzz_run_result_envelope, record/encode helpers, evidence-ref
builder, artifact-kind const) into a new
core::fuzz::result_envelope_persistence service, and add a
report_fuzz_result_envelope entry point that owns the report-time
write + persist + evidence flow. Commands (report, execution, inspect)
and tests now delegate to the shared core primitive.

Clears the fuzz/report.rs thin_command_adapter_violation (8 -> 7).
